+++ Tests/DependencyInjection/Fixtures/php/logging.php
@@ -1,0 +1,17 @@
+<?php
+
+/*
+ * This file is part of the colinodell\omnipay-bundle package.
+ *
+ * (c) 2015 Colin O'Dell <colinodell@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+$container->loadFromExtension('omnipay', [
+    'logging' => [
+        'enabled' => true,
+        'channel' => 'testchannel',
+    ],
+]);