+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## [Unreleased]
+
+## [0.2.0]
 ### Added
 
  - Added the ability to register new gateways via the container
@@ -12,4 +14,5 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
  - Initial release
 
-[Unreleased]: https://github.com/colinodell/omnipay-bundle/compare/0.1.0...HEAD
+[Unreleased]: https://github.com/colinodell/omnipay-bundle/compare/0.2.0...HEAD
+[0.2.0]: https://github.com/colinodell/omnipay-bundle/compare/0.1.0...0.2.0